Rename storyteller module to reporter to disambiguate from the storyteller dependency
