+++ src/lib.rs
@@ -15,16 +15,16 @@ use rust_releases::{semver, Channel, FetchResources, ReleaseIndex, RustChangelog
 use crate::check::RustupToolchainCheck;
 use crate::config::{Config, ModeIntent, OutputFormat, ReleaseSource};
 use crate::errors::{CargoMSRVError, TResult};
-use crate::storyteller::event::action::Action;
-use crate::storyteller::event::meta::Meta;
-use crate::storyteller::{Event, Reporter};
+use crate::reporter::event::action::Action;
+use crate::reporter::event::meta::Meta;
+use crate::reporter::{Event, Reporter};
 
 pub mod check;
 pub mod cli;
 pub mod config;
 pub mod errors;
 pub mod exit_code;
-pub mod storyteller;
+pub mod reporter;
 pub mod toolchain;
 
 pub(crate) mod command;
