+++ src/reporter/handler/discard_output_handler.rs
