+++ src/search_methods/linear.rs
@@ -4,9 +4,9 @@ use crate::check::Check;
 use crate::outcome::Outcome;
 use crate::{Config, TResult};
 // use crate::reporter::{write_failed_check, write_succeeded_check};
+use crate::reporter::Reporter;
 use crate::result::MinimalCompatibility;
 use crate::search_methods::FindMinimalCapableToolchain;
-use crate::storyteller::Reporter;
 use crate::toolchain::{OwnedToolchainSpec, ToolchainSpec};
 
 pub struct Linear<'runner, R: Check> {
