+++ src/subcommands.rs
@@ -5,7 +5,7 @@
 /// * Run `cargo msrv verify` on the CI, to verify the crates MSRV is acceptable.
 pub use {find::Find, list::List, set::Set, show::Show, verify::Verify};
 
-use crate::storyteller::Reporter;
+use crate::reporter::Reporter;
 use crate::{Config, TResult};
 
 pub(crate) mod find;
