+++ src/subcommands/find.rs
@@ -4,9 +4,9 @@ use crate::check::Check;
 use crate::config::{Config, SearchMethod};
 use crate::errors::{CargoMSRVError, TResult};
 use crate::releases::filter_releases;
+use crate::reporter::Reporter;
 use crate::result::MinimalCompatibility;
 use crate::search_methods::{Bisect, FindMinimalCapableToolchain, Linear};
-use crate::storyteller::Reporter;
 use crate::writers::toolchain_file::write_toolchain_file;
 use crate::writers::write_msrv::write_msrv;
 use crate::{Event, SubCommand};
